Agregado SP historialPrecioDe y backend

+++ SuperMercado/Product/HistorialPrecio.cs
@@ -2,7 +2,7 @@
 
 namespace SuperMercado.Product
 {
-    public record HistorialPrecio(Producto Producto, DateTime FechaHora, float PrecioUnitario)
+    public record HistorialPrecio(Producto Producto, DateTime FechaHora, decimal PrecioUnitario)
     {
         public bool Entre(DateTime inicio, DateTime fin)
             => FechaHora >= inicio && FechaHora <= fin;
